requirements.txt
flask==3.0.0
opencv-python
numpy
pillow
paho-mqtt
werkzeug
